shopper-app/src/main/java/com/qprogramming/shopper/app/security/oauth2/user/OAuth2UserInfoFactory.java
package com.qprogramming.shopper.app.security.oauth2.user;

import com.qprogramming.shopper.app.account.AccountType;
import com.qprogramming.shopper.app.exceptions.OAuth2AuthenticationProcessingException;

import java.util.Map;

public class OAuth2UserInfoFactory {

    public static OAuth2UserInfo getOAuth2UserInfo(String registrationId, Map<String, Object> attributes) {
        if (registrationId.equalsIgnoreCase(AccountType.GOOGLE.toString())) {
            return new GoogleOAuth2UserInfo(attributes);
        } else if (registrationId.equalsIgnoreCase(AccountType.FACEBOOK.toString())) {
            return new FacebookOAuth2UserInfo(attributes);
        } else {
            throw new OAuth2AuthenticationProcessingException("Sorry! Login with " + registrationId + " is not supported yet.");
        }
    }
}
